+++ ios/Classes/PrivacyScreenPlugin.h
@@ -1,4 +1,4 @@
 #import <Flutter/Flutter.h>
 
-@interface PrivacyScreenPlugin : NSObject<FlutterPlugin>
+@interface PrivacyScreenPlugin : NSObject <FlutterPlugin>
 @end