+++ lib/privacy_screen.dart
@@ -1,5 +1,3 @@
-library privacy_screen;
-
 export './src/privacy_helpers.dart';
 export './src/privacy_gate.dart';
 export './src/privacy_screen.dart';